[Added] Displaying Player Health On HUD.

+++ Assets/Scripts/Character/Player/PlayerHealth.cs
@@ -7,13 +7,13 @@ namespace VampireSurvivor
         public override void Damage(int damage)
         {
             base.Damage(damage);
-            EventManager.TriggerEvent(GameEvents.PlayerHealth, GetCurrentHealth);
+            EventManager.TriggerEvent(GameEvents.PlayerHealth, (float)GetCurrentHealth);
         }
 
         public override void Heal(int heal)
         {
             base.Damage(heal);
-            EventManager.TriggerEvent(GameEvents.PlayerHealth, GetCurrentHealth);
+            EventManager.TriggerEvent(GameEvents.PlayerHealth, (float)GetCurrentHealth);
         }
     }
 }
